feat(likx): v3.31.0 — PyPI distribution, clipboard paste, Wayland region selector, hotkey fixes
- Add PyPI entry point (`pip install likx`) with classifiers, keywords, build-system
- Add publish-pypi job to release workflow (trusted publishing)
- Clipboard image paste: Ctrl+V opens system clipboard image as new editor tab
- Wayland region selector: native slurp integration, GTK overlay fallback
- Fix global hotkey collisions: Ctrl+Shift+R/F/W → Super+Print, Super+Shift+S/W
- Fix effects.py memory leak: copy surface data before passing to Pixbuf.new_from_data
- Fix capture.py: use tempfile.mkstemp instead of predictable /tmp paths
- Fix capture.py: add None checks after Pixbuf.new_from_file
- Fix recorder.py: add text=True to subprocess calls (no more .decode())
- Replace print() with logging.warning() across effects, hotkeys modules
- Add timeout=5 to all gsettings subprocess.run calls
- Add gnome-screenshot -a fallback for Wayland region capture
- Rewrite CLAUDE.md with full architecture, shortcuts, conventions, gotchas

+++ likx/src/scroll_capture.py
@@ -218,8 +218,8 @@ class ScrollCaptureManager:
                 return False, None
 
             # Check if we're seeing the same content (no scroll happened)
-            if overlap >= height * 0.9:
-                return False, None  # Content didn't change, end reached
+            if overlap >= int(height * 0.9):
+                return False, None  # Nearly identical frames — end of scrollable content
 
             self.overlaps.append(overlap)
             self.frames.append(pixbuf)
